Clean up the repo before Gen3 conversion.
* Cleaning up new task files.
* Moving helper files into `utils` folder.
* Moving fluxCal to new task.
* Removing explicit `__all__` statements.
* Move task specific functions back to the individual task files, out of the `helpers` and `plotting` modules.
* Clean up and rename functions from camelCase to under_score.
* Remove `NoResults` exception for more generic because of problems/confusion with import differences with coming gen3.
* Convert the extraction qa to a thin pipeline task and fat task, like the others.

Remove `__init__` files for native namespacing. Update pyproject.toml file for proper version, install.
