--- python/pfs/drp/__init__.py
@@ -1,3 +1,0 @@
-import pkgutil
-
-__path__ = pkgutil.extend_path(__path__, __name__)
